:art: Refactor unit test
